requirements.txt
streamlit
pandas
scipy
google-generativeai
